Fix: Access sheets with UTF-8 characters in the name
Calling Workbook::sheets_by_name with a name containing an UTF-8 string, will now return the correct sheet.
